Evil: make behaviour of Y consistent.
In vim/evil, Y by default is a synonym for yy.  D and C are synonyms
for d$ and c$ respectively.  Y as a synonym for y$ is not only more
consistent, it is more useful: hitting $ is much more difficult than
repeating y.

./lisp/functions.el:

Added the function evil-yank-to-eol, which yanks from the point to end
of line.

./lisp/keys.el:

Added the keybinding Y -> evil-yank-to-eol.