Cleaned up some fancy-modeline stuff:
Significantly rewrote README.md:

Slightly expanded on the introduction.

Expanded on and corrected information on some features.

Wrote a quick guide to how it works and configuration.

Added a section on the unlicense

Cleaned up some of fancy-modeline.el:

Added comments to the setq mode-line-input and mode-line-active
sections noting that they don't actually *work*.

Changed the layout of the defface code so it actually fits inside 80
columns...