Sorting out colour settings:
Colour settings must be different for me between terminal and graphical
emacs.  The functions my-fix-fancy-modeline-colors and my-fix-colors set
various colors depending on whether the function is run in graphical or
terminal emacs.

The my-fix-colors command is bound to the <F3> key, meaning that it can
be run on-the-fly, enabling color-switching as I move between graphical
and terminal emacsclients.