Improve X selection handling
By default, emacs' handling of the clusterfuck that is copy-and-paste
under X11 is... weird.  In most applications, S-Ins and <Mouse-2> both
paste the X PRIMARY selection (i.e. the text selected with the mouse),
but in emacs only <Mouse-2> does this.  Having to move off the keyboard
onto the mouse in order to use the PRIMARY selection is very annoying.

Additionally, the function which is used to insert the primary selection
can ONLY be bound to mouse events.

We improve this by defining a function, 'my-yank-x-primary-selection,
which inserts the PRIMARY selection under the point.  This can actually
be bound to keys, and as such is immeasurably better.